+++ src/main/java/domain/CoordinateFactory.java
@@ -1,3 +1,5 @@
+package domain;
+
 import java.util.HashMap;
 import java.util.Map;
