+++ src/main/java/domain/Point.java
@@ -1,3 +1,5 @@
+package domain;
+
 import java.util.Objects;
 
 public class Point implements Comparable<Point> {
