+++ src/main/java/domain/Triangle.java
@@ -1,3 +1,5 @@
+package domain;
+
 import java.util.Arrays;
 import java.util.List;
