+++ src/main/java/domain/exception/InvalidCoordinateException.java
@@ -1,3 +1,5 @@
+package domain.exception;
+
 public class InvalidCoordinateException extends RuntimeException {
     private static final String MESSAGE = "유효하지 않는 좌표 값입니다.(좌표는 0이상 24이하의 정수입니다.)";
