+++ src/test/java/domain/PointTest.java
@@ -1,3 +1,5 @@
+package domain;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
